trading_bot/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
python-binance>=1.0.19
python-dotenv>=1.0
pydantic>=2.0
requests>=2.31
